+++ CHANGELOG.md
@@ -38,6 +38,9 @@ Additions
 - Added the ability for Dropdown to display an arrow texture based on its open state
 - Added the ability to specify Dropdown paragraph colors through style properties
 
+Improvements
+- Explicitly return the element type from Dropdown.AddElement overloads
+
 Fixes
 - Fixed tooltips not being bounded correctly for viewports that don't start at the origin
 - Fixed a stack overflow exception when a panel's children have just enough height to cause a scroll bar to appear